pulsar-testing/requirements.txt
pulsar-client==3.0.0
